Software/Python/requirements.txt
numpy
smbus-cffi
RPi.GPIO
pyserial
scipy
